Lock message data field sizes to 4GB

+++ max/m_change.c
@@ -56,7 +56,7 @@ int Msg_Change(void)
 
 
   if ((msgh=MsgOpenMsg(sq, MOPEN_READ, lmsg))==NULL ||
-      MsgReadMsg(msgh, &msg, 0L, 0L, NULL, 0L, NULL)==(unsigned long)-1L ||
+      MsgReadMsg(msgh, &msg, 0L, 0L, NULL, 0L, NULL)==(uint32_t)-1 ||
       ! (eqstri(msg.from, usr.name) || eqstri(msg.from, usr.alias) ||
          mailflag(CFLAGM_ATTRANY)))
   {
